requirements.txt
# Minimal requirements
pillow>=6.0.0
pywin32; platform_system == "Windows"
comtypes; platform_system == "Windows"